LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/timeline-offset-in-keyframe-change-timeline.tentative-expected.txt


FAIL getKeyframes with timeline-offsets assert_equals: properties on ComputedKeyframe #0:  should match expected "composite,computedOffset,easing,marginLeft,offset,opacity" but got "composite,computedOffset,easing,marginLeft,marginRight,offset,opacity"
